python-ocr-service/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
opencv-python-headless
pillow
